Fix issue with multiple clicks
When clicking multiple times (like a double click), BasicRobot
would generate a "move mouse" event between each click, even if it
was to move the mouse to the place it already was at. On Mac this
was interpreted as several single clicks, instead of a multi-click.
This removes the unnecessary move events.

Change-Id: Ieda948bfa1dc71ce2721a8b0199876278afaf26b