Dead lock when Dialog/FrameFixture's are created with a null target
During Object construction, DialogFixture/FrameFixture make a call to
robotWithCurrentAwtHierarchy(), which acquires the screen lock, and
only after that check for the validity of the constructor parameters.
If a parameter turns out to be invalid (ie null), an exception will
be thrown and the Object does not finish initialisation, leaving no
way to reset the screen lock.
This CL changes the code to check the parameters before attempting to
acquire the screen lock.

Test: FrameFixture_constructor_withFrame_Test
Change-Id: Id8b6326e2a5cd459a84996679bad9c7dd5e5b374